Refactor featured carousel components
