--- ui-service/src/main/frontend/hotel-booking-ui/src/components/FeaturedBundle/types.ts
@@ -1,6 +1,0 @@
-export interface FeaturedBundle {
-  destination: string;
-  dates: string;
-  price: number;
-  imgSrc: string;
-}
