+++ ui-service/src/main/frontend/hotel-booking-ui/src/components/FeaturedCarouselComponent/featuredPropertiesMockData.ts
